правки

+++ src/bot/callback.py
@@ -489,6 +489,13 @@ def chest_callback(call: CallbackQuery):
         bot.delete_message(call.message.chat.id, call.message.id)
         if call.message.reply_to_message:
             bot.reply_to(call.message.reply_to_message, "*Ушел от сундука*")
+    elif data[1] == "bag":
+        markup = InlineMarkup.bag(user)
+        text = "Инвентарь"
+
+        bot.edit_message_text(
+            text, call.message.chat.id, call.message.id, reply_markup=markup
+        )
 
 
 @bot.callback_query_handler(lambda c: c.data.startswith("guide"))
@@ -822,7 +829,8 @@ def delate_state_callback(call: CallbackQuery):
         return
 
     bot.delete_state(call.from_user.id, call.message.chat.id)
-    bot.delete_message(call.message.chat.id, call.message.id)
+    if call.message.id:
+        bot.delete_message(call.message.chat.id, call.message.id)
     bot.answer_callback_query(call.id, "Отменил")
